add CRUD operation to for mealapp.

+++ mealapp/views.py
@@ -1,49 +1,74 @@
-from django.shortcuts import get_object_or_404
-from rest_framework import status
+from django.conf import settings
+from django.http import JsonResponse
+from django.views.decorators.csrf import csrf_exempt
+from rest_framework.parsers import JSONParser
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated
-from rest_framework.response import Response
-from .models import Meal, FoodItem
-from .serializers import MealSerializer, FoodItemSerializer
+from .models import Meal
+from .serializers import MealSerializer
 
-# Get all meals or create a new meal
+# Get all meals for all users or create a new meal
+@csrf_exempt
 @api_view(['GET', 'POST'])
-@permission_classes([IsAuthenticated])
-def meal_list(request):
-    if request.method == 'GET':
+def meals_list(request):
+    
+    if request.method == 'GET':       
         meals = Meal.objects.all()
         serializer = MealSerializer(meals, many=True)
-        return Response(serializer.data)
+        return JsonResponse(serializer.data, safe=False)
     elif request.method == 'POST':
-        serializer = MealSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save(user=request.user)
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        data = JSONParser().parse(request)
+        data['user'] = request.user.id
+        
+        try:
+            meal = Meal.objects.get(user=request.user, selectedDate=data['selectedDate'])
+            serializer = MealSerializer(meal)
+            return JsonResponse({'error': f"Meal already created for {data['selectedDate']}"})
+        except Meal.DoesNotExist:
+            pass
 
-# Get, update, or delete a single meal
+        serializer = MealSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save(user=request.user)  # pass the authenticated user
+            return JsonResponse(serializer.data, status=201)
+        return JsonResponse(serializer.errors, status=400)
+
+
+# Get, update or delete a single meal
+@csrf_exempt
 @api_view(['GET', 'PUT', 'DELETE'])
-@permission_classes([IsAuthenticated])
-def meal_detail(request, pk):
-    meal = get_object_or_404(Meal, pk=meal_id)
+def meal_detail(request, id):
+    try:
+        meal = Meal.objects.get(id=id)
+    except Meal.DoesNotExist:
+        return JsonResponse({'message': 'The meal does not exist'}, status=404)
+
+    if meal.user != request.user:
+        return JsonResponse({'message': 'Unauthorized'}, status=401)
+
     if request.method == 'GET':
         serializer = MealSerializer(meal)
-        return Response(serializer.data)
+        return JsonResponse(serializer.data)
+
     elif request.method == 'PUT':
-        serializer = MealSerializer(instance=meal, data=request.data)
+        data = JSONParser().parse(request)
+        serializer = MealSerializer(meal, data=data)
         if serializer.is_valid():
-            serializer.save(user=request.user)
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            serializer.save()
+            return JsonResponse(serializer.data)
+        return JsonResponse(serializer.errors, status=400)
+
     elif request.method == 'DELETE':
         meal.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        return JsonResponse({'message': 'Meal was deleted successfully!'}, status=204)
 
+# Get all meals for a single user
+@csrf_exempt
+@api_view(['GET'])
+def user_meals(request, user_id):
+    if request.user.id != user_id:
+        return JsonResponse({'message': 'Unauthorized'}, status=401)
 
-# get all meal for single user
-@api_view(["GET"])
-@permission_classes([IsAuthenticated])
-def getMealsByUserId(request, pk):
-    meals = Meal.objects.filter(user_id=pk)
+    meals = Meal.objects.filter(user=user_id)
     serializer = MealSerializer(meals, many=True)
-    return Response(serializer.data)
+    return JsonResponse(serializer.data, safe=False)
